feat: Wire active_zones to HTTP search/retrieve API endpoints
- Moved `parse_zones_from_prompt` to `src/memory/schema.rs` to share it across the codebase.
- Added `active_zones` field to `MultiLayerRetrieveRequest`, `V1SearchRequest`, and `SearchPayload`.
- Wired `active_zones` in `build_multi_layer_retrieve_response`, `v1_memories_search`, and `search_handler` to enable explicit zone selection and fallback to keyword-based detection.
- Updated `MemoryQueryFilters` usage in API handlers to include the specified or detected zones.
