+++ src/context/orchestrator.rs
@@ -101,7 +101,7 @@ impl Orchestrator {
         documents: &[ContextDocument],
     ) -> ExecutionPlan {
         let level = self.classifier.classify(prompt);
-        let active_zones = parse_zones_from_prompt(prompt);
+        let active_zones = crate::memory::schema::parse_zones_from_prompt(prompt);
         let session_documents: Vec<_> = documents
             .iter()
             .filter(|document| document.session_id == session_id)
@@ -378,32 +378,6 @@ struct PlanConfig {
     max_tokens: usize,
     include_tool_calls: bool,
     include_metadata: bool,
-}
-
-fn parse_zones_from_prompt(prompt: &str) -> Vec<ContextZone> {
-    let lowered = prompt.to_lowercase();
-    let mut zones = Vec::new();
-
-    if lowered.contains("detalle") || lowered.contains("atómico") || lowered.contains("atomic") {
-        zones.push(ContextZone::Atomic);
-    }
-    if lowered.contains("resumen") || lowered.contains("cluster") || lowered.contains("agrupar") {
-        zones.push(ContextZone::Cluster);
-    }
-    if lowered.contains("general") || lowered.contains("global") || lowered.contains("estrategia") {
-        zones.push(ContextZone::Global);
-    }
-    if lowered.contains("relación") || lowered.contains("vínculo") || lowered.contains("grafo") || lowered.contains("belief") {
-        zones.push(ContextZone::Relational);
-    }
-
-    if zones.is_empty() {
-        // Default zones based on implicit intent if none specified
-        zones.push(ContextZone::Atomic);
-        zones.push(ContextZone::Cluster);
-    }
-
-    zones
 }
 
 fn build_query(prompt: &str, level: ContextLevel, hook: HookKind) -> String {
